refactor - Moved Coordinate class to linear plot data package.
