+++ charts/src/main/java/dev/anirban/charts/linear/data/BasicDataStrategy.kt
@@ -6,7 +6,6 @@ import android.graphics.Typeface
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.drawscope.DrawScope
 import androidx.compose.ui.unit.sp
-import dev.anirban.charts.util.Coordinate
 
 
 /**
